android/foundation: Move RecyclerView-related presentation components
Move RecyclerView-related classes into a dedicated package. With the
planned new presenter, we would need to extract the nested classes from
the old one. This new package shall then be their home.
