+++ lib-kmm-foundation/src/androidMain/kotlin/com/hadisatrio/libs/android/foundation/widget/recyclerview/NaiveViewFactory.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2022 Hadi Satrio
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.hadisatrio.libs.android.foundation.widget.recyclerview
+
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView.LayoutParams.MATCH_PARENT
+import androidx.recyclerview.widget.RecyclerView.LayoutParams.WRAP_CONTENT
+import kotlin.math.roundToInt
+
+internal object NaiveViewFactory : ViewFactory {
+
+    override fun create(parent: ViewGroup, viewType: Int): View {
+        return TextView(parent.context).apply {
+            val density = parent.resources.displayMetrics.density
+            val itemPadding = (density * RecyclerViewPresenter.DEFAULT_ITEM_PADDING).roundToInt()
+            layoutParams = RecyclerView.LayoutParams(MATCH_PARENT, WRAP_CONTENT).apply {
+                setMargins(itemPadding, itemPadding, itemPadding, itemPadding)
+            }
+        }
+    }
+}
